[sw/husky] Make batches of 1 possible
For the NTT subfunction of ML-DSA the max batch size
is 1 since it takes 8791 cycles to execute.
To still be able to run fvsr tests it is much easier to
allow for batches of size 1 than to adapt the whole capture
script along with the ibex c code and probably even more.